Eliminate the "session manager" process from slurmd.
Now the slurmd job-step manager, which runs as root, directly forks
and execs all of the tasks in the job step.  The tasks are all placed
in the same process group.

To be more compatible with the new task launch layout, the proctrack API
was slightly modified.  The slurm_create_container function now tasks a
slurmd_job_t pointer as its only parameter.  Each proctrack plugin can
then decide which information it needs to look at to register the tasks.