Merge -r5338:5560 from stdio-devel branch onto trunk.
This merge is a rewrite of slurm's task standard I/O handling.
srun will now have just a single TCP stream to each node of a job-step.
srun and slurmd comminicate over the TCP stream using a simple messaging
protocol.

Both srun and slurmd now use a completely event driven I/O engine based on
the eio.c event I/O engine.

Be warned that the new I/O support in this merge is not quite complete.  The
core I/O protocol is functional, but it is still missing reattach support
and it needs extensive testing and debugging.