Fixes a flaw in client eio_obj_t setup.  There are differences in how the initial client eio object and later "srun --attach" objects are handled due to the caching.
Modifications to the eio engine to move the List of eio_obj_t into the
opaque eio_handle_t.  New eio calls "eio_new_initial_obj" and
"eio_new_obj", both of which add an eio_obj_t to the eio_handle_t's internal
object List.  However, only the later function is safe to call when
the eio mainloop is running.

Also caught in this commit is some initial code to implement the RPCs
REQUEST_TERMINATE_* and REQUEST_SIGNAL_*.