Fix segfault in the slurmstepd on unexpected input.
Srun and/or slurmctld are doing the wrong thing on "srun -N1-4 -n3".  They
are launching 3 tasks on 4 nodes, with the 4th node getting a launch RPC
for zero tasks.  This patch fixes a slurmstepd segfault when the task count
is zero, but the controller should not be giving srun 4 nodes for only
three tasks.