Vastly improve slaunch's signal handling abilities.
slaunch will now cancel a job step on SIGINT, SIGHUP, and SIGTERM.  If
a signal it caught, it will signal the step to terminate and wait up
to 10 seconds for all remote nodes to report that the tasks have exited.

After 10 seconds, it will give up waiting and exit, regardless of the
state of the job step.