Add UnkillableStepProgram and UnkillableStepTimeout variables to the slurm.conf.  This feature was requested by Py for AIX.
There are times when user processes will become unkillable because of
dirty NFS pages that AIX is too stupid to realize need flushing.  Py wrote
and program/script that can temprorily disable the cache which allows
the user processes to die.

This feature will allow that external program to be executed only when
the processes in a job step refuse to die for some period of time.