In sbatch I found that the --propagate user option was doing nothing when there was a PropagateResourceLimits* option in slurm.conf. It looks like the conf file was being read *after* the user options for rlimits propagation were being set, and thus the user request was being overridden.
Furthermore, it appears that the loop in _set_rlimit_env()
does not check for whether the current rlimit has even been
chosen to be propagated. Instead it sets all limit information
in the environment regardless of the propagate_flag value.

This patch moves the parsing of the opt.propagate flag in
sbatch to after the conf file is read, so that user options
get precedence. Furthermore, only rlimits that have been
chosen for propagation are actually propagated.

Mark Grondona