Ensure slurmctld adopts SlurmUser GID as well as UID.
This patch abstracts the establishment of slurm user credentials
in slurmctld when SlurmUser != root. It adds a call to setgid()
to adopt the SLURM user's primary group, instead of running with
whatever primary group was in effect when invoking the slurmctld
executable. Also, errors in establishing user credentials are
fatal errors, instead of just warnings.

This is a minor security enhancement.