Thread-safety fixes for proctrack/lua
Multiple threads of slurmd/slurmstepd should not be allowed to call into
the proctrack/lua script simultaneously. This was causing segfaults
under slurm_container_signal as reported by Aaron Knister.
The simplest fix here is a global proctrack/lua lock (as suggested by
Aaron), that only allows one thread at a time to be in proctrack/lua
callbacks. This patch implements the global lua_lock.

Thanks to Aaron Knister for the original version of this patch.

Reported-by: Aaron Knister <aaron.knister@gmail.com>