job_submit/lua: remove unused reference to 'posix' module
The sample code does not refer to the 'posix' Lua module. However, the
absence of the module is sufficient to keep slurmctld from starting, which is
confusing when trying to learn the capabilities of the job_submit/lua module.

Removing the inclusion does not hurt the code and makes it easier to get
started.  Patch from Gerrit.