Patch from Gerrit 22_Cray-slurmctld-SlurmUser-must_be-root.diff
Cray: SlurmUser needs to be 'root'

This disallows a SlurmUser other than root. The reason is that slurmctld forks
processes talking to ALPS/BASIL. Doing this requires root permissions.

The following rules apply:
 * the RESERVE method requires that 
   - either the effective UID of the calling process resolves to that of the
     'user_name' attribute of the ReserveParamArray or
   - the effective UID is 0 (super-user), i.e.
   - only root can make reservations for other users;
 * the RELEASE method requires root permissions in all cases (slurmctld uses
   this method on each scheduling cycle to clean up vestigial ALPS reservations).

To avoid entanglements with sudo, this patch simply throws a fatal() message
if the SlurmUser is not root. (An alternative would be to issue an error 
message, but to continue processing.)