Patches from Gerrit: 24_Cray-deprecate-HAVE_CRAY.diff and
various files: deprecate HAVE_CRAY

This 
 * replaces HAVE_CRAY (which is an untested configure option) 
 * with HAVE_NATIVE_CRAY which tests for Cray-specific files and
 * deprecates the apbasil location (which has been /usr/bin/apbasil since XT3).

This distinction is used to disable code on systems that do not support Cray
ALPS interaction.

25_Cray-test-for-native-cray-system.diff
Cray: test for native system

This changes the autoconf logic so that
 * HAVE_CRAY is defined to enable basic Cray XT/XE support;
 * HAVE_NATIVE_CRAY is defined on native Cray XT/XE systems only.

The distinction is necessary due to the user of multi-cluster systems, which
may need to load the select/cray plugin, but which are not native Cray systems
(such as e.g. the slurmdbd host).

In particular, when building with support for remote execution of Cray jobs,
it would be wrong to set FRONT_END mode locally on the clients. Hence only
set this when building on a native Cray system.