-- Set Cray node order based upon ALPS_NIDORDER configuration.     03_Cray-BASIL-node-ranking.diff select/cray: perform node ranking
This supplies the select function-pointer to request a reordering of nodes based
on the current Cray node ordering. 

The Cray node ordering is set internally via the ALPS_NIDORDER configuration 
variables that controls the way ALPS considers nodes.

This ordering in turn determines the order of nodes as the appear subsequently 
in the Inventory output. The present patch exploits this fact and uses an
auto-incrementing number to reflect the node ranking (counting is reversed 
since the parser returns the nodes in stack/LIFO order).

The node ranking is performed on slurmctld (re-)configuration, hence the tests
are more stringent: exit if Inventory fails (this condition is extremely rare)
and if no nodes are powered up (also a condition that can be cured by restarting
slurmctld only when the system is ready).