Do a call of select_g_reconfigure() on Cray systems immediately before attempting to schedule jobs. 04_Cray-INVENTORY-directly-before-schedule.diff
select/cray: do the inventory immediately before each schedule

This puts the Basil inventory immediately before each (backfill) schedule. 

Having considered multiple alternatives, this is the most robust and least
wasteful solution. The reason is that ALPS keeps internal node state, which
can be changed
 * by the administrator (xtprocadmin),
 * by the node health checker programs (setting some nodes into 'suspect'),
 * by ALPS itself.

Tracking this periodically, e.g. every HealthCheckInterval, may mean to miss
some state changes. The result would not be a crash, but a subsequently
failed ALPS reservation, which would require to undo some of the slurm state.

FIXME: since we are not using this, we have not yet considered wiki/wiki2.
       Possible places to update these are:
       - run periodic checkes every HealthCheckInterval (sub-optimal),
       - change plugins/sched/wiki{,2}/get_nodes.c