node_select: revert a change which broke in compatibility mode
If this test is performed on a non-Cray system which tries to talk to 
a remote Cray system, it fails -- which it should not.

ela1:1 ~>echo $SLURM_CLUSTERS
palu
ela1:0 ~>squeue
squeue: fatal: Requested SelectType=select/cray in slurm.conf, but not running on a cray system.  If looking to emulate a Cray system use --enable-cray-emulation.

02_node_select_test.diff