read_config: make sure that local system has minimum Cray support
This is a global compatibility test to ensure that any (remote) host trying to talk
to a cluster using select/cray meets the minimum requirements of supporting the
required Cray data structures and hooks.

As per previous patches, it may be possible to factor this out, but at this stage
is working code.

06_read_config--test-for-select-cray.diff