salloc: disable --no-shell mode
Since "aprun" is used on Cray instead of srun, the --no-shell option does not
make any difference: with or without this option, the ALPS reservation is made, 
and since it is confirmed using the SID of the current shell, aprun will run 
even if the BASIL_RESERVATION_ID is not set.

NB: the patch aborts with an error message. If deciding to turn this into a
    warning, and continue processing, opt.no_shell should be disabled, since
    otherwise interactive mode (and thus job control) is disabled.