select/cray: more carefully test for NULL job pointer
This audits the select/cray code so that it does not accidentally dereference a NULL job_ptr.
This instance happens once, upon restart of slurmctld (detailed description below). 
Similar checks are also in place in other select plugins, in any case it is better to check this.
Almost all cases use xassert(), the only exception is p_job_fini(), which assumes NULL means
there is nothing to be finalized.