scontrol: set uid when releasing a job
This fix is related to an earlier one and was observed when trying to 'scontrol release'
a job previously submitted via 'sbatch --hold' by the same user.

Within the job_submit/lua plugin, the user gets automatically assigned a partition. So,
even if no submitter uid checks are usually expected, it can happen in the process of
releasing a job, that a part_check is performed.

In this case, the error message was

[2011-03-30T18:37:17] _part_access_check: uid 4294967294 access to partition usup denied, bad group
[2011-03-30T18:37:17] error: _slurm_rpc_update_job JobId=12856 uid=21215: User's group not permitted to use this partition

and like before (in scontrol_update_job()), was fixed by supplying the UID of the requesting user.