slurmctld: case of authorized operator releasing user hold
This patch avoids that the priority is not recalculated on 'scontrol release',
which  happens when an authorized operator releases a job, or if the job is
released via e.g. the job_submit plugin.

The patch reorders the tests in update_job() to 
 * test first if the job has been held by the user and, only if not,
 * test whether an authorized operator changed the priority or
   the updated priority is being reduced.

Due to earlier permission checks, we have either
 * job_ptr->user_id == uid or 
 * authorized,
where in both cases the release-user-hold operation is authorized.