api: remove unreferenced and undocumented function
This removes a function "slurm_pack_msg_no_header" which is nowhere referenced
in the src tree and which also is not listed in any of the slurm manpages.

As far as I understand the documentation, each slurm message needs to have a
header, it could thus be that this function is from very old or initial code.