select/cray: unconditionally release reservations
This increases robustness in releasing ALPS reservations. Previously
the reservation was only released through
 * select_g_job_fini() for interactive (salloc) sessions;
 * batch_finish() by slurmstepd for batch sessions.

This introduces a single point of failure for batch jobs, since a failure
of batch_finish() would mean that the reservation could only be released
much later, through the detection of orphaned ALPS reservations in
basil_inventory().

For batch jobs that terminate normally this means that the RELEASE method is
called twice: first in job_complete(), and then in batch_finish(). The Basil
1.2 design document by Ben Landsteiner (dated 15 Feb 2011) suggests in section
3.3.5 repeated calls of RELEASE as one possible way of improving the response
of the RELEASE method. There will be additional "entry not found" messages in
the apschedMMDD logs, but (due to the preceding patch) not in the SLURM logs.

For jobs that have to be terminated (e.g. job_timed_out, job_requeue, job_fail),
this patch will mean that the RELEASE is called much sooner and thus is 
expected to improve efficiency.

For interactive salloc sessions that are cancelled via scancel, there is now
no longer a warning message about the no longer existing ALPS reservation
(since the release happens first through select_p_job_signal and then through
 job_complete -> deallocate_nodes -> select_p_job_fini).