slurmctld: allow job_submit plugin to modify/set the priority/nice values
This allows the job_submit plugin to directly set priority values. If it
assigns a priority value different from 0 and NO_VAL, the priority is marked
as "fixed" via job_ptr->direct_set_prio.

To enable this, the permission check for directly set priority is now done
before calling the job_submit plugin, which in addition also allows to 
influence the nice value of the job via the plugin.