select/cray: add data structure support for confirmation cookie
This extends the Cray-specific select_jobinfo struct with a confirmation cookie
field, which is to be used by later patches to store the session SID or PAGG
container ID.

There is a slight incompatibility with regard to pack/unpack, due to the new
confirm_cookie format. Since not many Cray installations exist yet, I would
like to suggest to not do an extra bump of the API version.