select/cray: use pagg ID to confirm reservations
This uses the SGI container process aggregate ID to confirm the job reservation.

It falls back to using the alloc_sid in case of failure. This fallback
should be considered really only as a last resort, since it is known that
session IDs are not unique across multiple login nodes and therefore the
confirmation of ALPS reservations will fail whenever there is a SID collision
(the likelihood increases with system size).