select/cray: warn user of wrong ProctrackType
To work properly, select/cray requires proctrack/sgi_job. In fact, due to the
way the container functions are called by the slurmstepd, it will only work
properly with this plugin.

I have considered alternatives, such as falling back to using the SID to
confirm the allocation. But this attempt to support configuration errors
creates other problems, such as less stringent error checking.