salloc: support for Cray container IDs
This adds detection and use of SGI process aggregate job container IDs for
salloc interactive sessions.

The preferred and documented way to support this on a Cray system is by
enabling the provided pam_job.so via /etc/pam.d/common-session.

There is a header dependency on job.h. This file depends on the optional
cray-libjob-devel package, which installs into /opt/cray/job/<version>.

This package is however not always installed or may not be up-to-date.
Hence the patch "cheats" by duplicating the known prototype of job_getjid().