select/cray: document the differences between SLURM/ALPS memory description
This adds user documentation on how memory is specified by ALPS (per-PE share) and SLURM
(per-node or per-CPU units), clarifying the correspondence.
Since it is a correspondence, but not a direct equivalence, the patch also removes the
PBS-script directive support for 'mppmem', since otherwise it would lead to unexpected
behaviour. From Gerrit Renker, CSCS.