Improve cray/srun/aprun wrapper task layout
Improve how the cray srun/aprun wrapper handles the ntasks and nnodes
options to better distribute tasks over the allocated nodes. The
support for these options is imperfect for resource allocations in
which the number of tasks per node is not uniform, but that can not
be properly handled due to differences between srun and aprun.