Add --enable-salloc-kill-cmd build option
Add an configure option of --enable-salloc-kill-cmd which will
cause the salloc command to signal its child processes when the
job ends. Job signalling will be the default for Cray systems.
Jobs will not be signalled by default on other systems. SIGHUP
will be used for interactive jobs and SIGTERM will be used for
other jobs.