Modify man2html calling for clean build in src dir
Modify the man2html calling sequence for a clean build of SLURM from the
source directory. The old code would copy the header.txt and footer.txt
files from "srcdir" to "." and generate an error if they were the same
files. The new code just passes the pathnames of the files into the
man2html python script