Add gang flag to select plugin job suspend/resume APIs
Add flag to the select APIs for job suspend/resume indicating if the action
is for gang scheduling or an explicit job suspend/resume by the user. Only
an explicit job suspend/resume will reset the job's priority and make
resources exclusively held by the job available to other jobs. This change
is also needed for Cray systems with ALPS.